move components.py to components folder
