+++ src/python/impactx/dashboard/__init__.py
@@ -11,7 +11,7 @@ from .Input.csrConfiguration.csrMain import csrConfiguration
 from .Input.distributionParameters.distributionMain import DistributionParameters
 from .Input.inputParameters.inputMain import InputParameters
 from .Input.latticeConfiguration.latticeMain import LatticeConfiguration
-from .Input.components import NavigationComponents
+from .Input.components.components import NavigationComponents
 from .Input.space_charge_configuration.spaceChargeMain import SpaceChargeConfiguration
 
 from .jupyterApplication import JupyterMainApplication as JupyterApp
